requirements.txt
django==5.0.7
django-environ==0.11.2
pillow==10.4.0
djangorestframework==3.15.2
PyJWT==2.8.0
